refactor(cloud): put the provider-specific components under aws/
Every component here extends sst.aws.* — Api, Cron, Function, Queue, Storage,
Topic, and now ObjectStorage — while the directory said nothing about it, which
is provider vocabulary leaking into a neutral layer one level up from where the
design already forbids it.

They now sit in src/sst/aws/. What stays outside is what is genuinely not
provider-specific: App, Stack, Linkable, LinkedEnvironment, naming, and
fromManifest, which is the kind → component registry and therefore the seam a
second provider extends.

The barrel is unchanged, so @geekmidas/cloud/sst exports exactly what it did and
no consumer notices. GCP is not being built; this is only so that adding it
later is a directory beside this one rather than a restructure.
