+++ packages/cloud/src/sst/index.ts
@@ -1,3 +1,4 @@
+export { App, type AppProps, type StageValues } from './App';
 export {
 	Api,
 	type ApiAuthorizers,
@@ -6,17 +7,24 @@ export {
 	type JwtAuthorizer,
 	type LambdaAuthorizer,
 	type Route,
-} from './Api';
-export { App, type AppProps, type StageValues } from './App';
+} from './aws/Api';
 export {
 	Cron,
 	type CronExpression,
 	type CronProps,
 	type CronRate,
 	type CronSchedule,
-} from './Cron';
+} from './aws/Cron';
+export { Function, type FunctionProps } from './aws/Function';
+export {
+	ObjectStorage,
+	type ObjectStorageProps,
+} from './aws/ObjectStorage';
+export { Queue, type QueueProps } from './aws/Queue';
+/** @deprecated renamed to `ObjectStorage`, matching the construct that declares it. */
+export { Storage, type StorageProps } from './aws/Storage';
+export { Topic, type TopicProps } from './aws/Topic';
 export { ProvidesMismatch, UnknownDeclarationKind } from './errors';
-export { Function, type FunctionProps } from './Function';
 export {
 	assertProvides,
 	type ComponentOverrides,
@@ -26,12 +34,4 @@ export {
 	provisionerFor,
 } from './fromManifest';
 export { type GkmLinkable, ResourceType } from './Linkable';
-export {
-	ObjectStorage,
-	type ObjectStorageProps,
-} from './ObjectStorage';
-export { Queue, type QueueProps } from './Queue';
 export { Stack, type StackType } from './Stack';
-/** @deprecated renamed to `ObjectStorage`, matching the construct that declares it. */
-export { Storage, type StorageProps } from './Storage';
-export { Topic, type TopicProps } from './Topic';
